Split htmlsource and jsonsource
Add json source tests
