+++ scraper/src/source/seloger.ts
@@ -1,4 +1,4 @@
-import { HTMLSource } from '../types/source'
+import { HTMLSource } from '../types/html-source'
 import { getSize } from '../helper/size'
 
 export default class Seloger extends HTMLSource {
